requirements.txt
Django==3.2
django-environ==0.4.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
drf-yasg==1.20.0
celery==5.2.1
django-celery-beat==2.2.1